requirements.txt
torch
torchvision
numpy
scikit-learn
matplotlib
seaborn
pandas
streamlit
pillow